pyproject.toml
[project]
name = "mechanicalmanager"
version = "0.1.0"
description = ""
authors = [
    {name = "mateuslinhares16",email = "mateuslinhares16@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-jazzmin (>=3.0.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-crispy-forms (>=2.4,<3.0)",
    "crispy-bootstrap5 (>=2025.4,<2026.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
